+++ pennylane/tape/tapes/__init__.py
@@ -15,9 +15,7 @@
 This subpackage contains various quantum tapes, which track, queue,
 validate, execute, and differentiate quantum circuits.
 """
-from .circuit_graph import NewCircuitGraph
 from .tape import QuantumTape
 from .qubit_param_shift import QubitParamShiftTape
 from .cv_param_shift import CVParamShiftTape
-from .rev import ReversibleTape
-from .qnode import QNode, qnode
+from .reversible import ReversibleTape
