+++ tests/tape/interfaces/test_qnode_tf.py
@@ -19,8 +19,8 @@ tf = pytest.importorskip("tensorflow", minversion="2.1")
 import numpy as np
 
 import pennylane as qml
-from pennylane.beta.tapes import QuantumTape, qnode, QNode
-from pennylane.beta.queuing import expval, var, sample, probs
+from pennylane.tape import QuantumTape, qnode, QNode
+from pennylane.tape.measure import expval, var, sample, probs
 
 
 @pytest.mark.parametrize(
