+++ tests/tape/interfaces/test_tape_tf.py
@@ -19,9 +19,9 @@ tf = pytest.importorskip("tensorflow", minversion="2.1")
 import numpy as np
 
 import pennylane as qml
-from pennylane.beta.tapes import QuantumTape
-from pennylane.beta.queuing import expval, var, sample, probs
-from pennylane.beta.interfaces.tf import TFInterface
+from pennylane.tape import QuantumTape
+from pennylane.tape.measure import expval, var, sample, probs
+from pennylane.tape.interfaces.tf import TFInterface
 
 
 class TestTFQuantumTape:
